+++ spec/Request/ContaoPageContextInitializerSpec.php
@@ -1,43 +1,29 @@
 <?php
 
-/**
- * Contao Page Context
- *
- * @package    contao-page-context
- * @author     David Molineus <david.molineus@netzmacht.de>
- * @copyright  2018 netzmacht David Molineus.
- * @license    LGPL-3.0 https://github.com/netzmacht/contao-page-context/blob/master/LICENSE
- * @filesource
- */
-
 declare(strict_types=1);
 
 namespace spec\Netzmacht\Contao\PageContext\Request;
 
-use Contao\CoreBundle\Framework\Adapter;
-use Contao\CoreBundle\Framework\ContaoFrameworkInterface;
+use Contao\CoreBundle\Framework\ContaoFramework;
 use Contao\CoreBundle\Image\PictureFactoryInterface;
 use Netzmacht\Contao\PageContext\Request\ContaoPageContextInitializer;
-use Netzmacht\Contao\Toolkit\Callback\Invoker;
 use Netzmacht\Contao\Toolkit\Data\Model\RepositoryManager;
 use PhpSpec\ObjectBehavior;
 use Psr\Log\LoggerInterface;
-use Symfony\Component\Translation\TranslatorInterface;
+use Symfony\Contracts\Translation\LocaleAwareInterface;
 
 final class ContaoPageContextInitializerSpec extends ObjectBehavior
 {
     public function let(
-        TranslatorInterface $translator,
-        ContaoFrameworkInterface $framework,
+        LocaleAwareInterface $translator,
+        ContaoFramework $framework,
         PictureFactoryInterface $pictureFactory,
         RepositoryManager $repositoryManager,
-        LoggerInterface $logger,
-        Adapter $systemAdapter
+        LoggerInterface $logger
     ): void {
         $this->beConstructedWith(
             $translator,
             $framework,
-            new Invoker($systemAdapter),
             $pictureFactory,
             $repositoryManager,
             $logger