+++ spec/Security/PageContextVoterSpec.php
@@ -1,15 +1,5 @@
 <?php
 
-/**
- * Contao Page Context
- *
- * @package    contao-page-context
- * @author     David Molineus <david.molineus@netzmacht.de>
- * @copyright  2018 netzmacht David Molineus.
- * @license    LGPL-3.0 https://github.com/netzmacht/contao-page-context/blob/master/LICENSE
- * @filesource
- */
-
 declare(strict_types=1);
 
 namespace spec\Netzmacht\Contao\PageContext\Security;
@@ -25,9 +15,7 @@ use Symfony\Component\Security\Core\Authorization\Voter\VoterInterface;
 
 final class PageContextVoterSpec extends ObjectBehavior
 {
-    /**
-     * @var PageContext
-     */
+    /** @var PageContext */
     private $pageContext;
 
     public function let(