+++ src/DependencyInjection/Compiler/PageIdDeterminatorPass.php
@@ -1,15 +1,5 @@
 <?php
 
-/**
- * Contao Page Context
- *
- * @package    contao-page-context
- * @author     David Molineus <david.molineus@netzmacht.de>
- * @copyright  2018 netzmacht David Molineus.
- * @license    LGPL-3.0 https://github.com/netzmacht/contao-page-context/blob/master/LICENSE
- * @filesource
- */
-
 declare(strict_types=1);
 
 namespace Netzmacht\Contao\PageContext\DependencyInjection\Compiler;
@@ -19,19 +9,15 @@ use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\Compiler\PriorityTaggedServiceTrait;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 
-/**
- * Class PageIdDeterminatorPass
- */
+use function array_merge;
+
 final class PageIdDeterminatorPass implements CompilerPassInterface
 {
     use PriorityTaggedServiceTrait;
 
-    /**
-     * {@inheritdoc}
-     */
     public function process(ContainerBuilder $container): void
     {
-        if (!$container->hasDefinition(PageIdDeterminator::class)) {
+        if (! $container->hasDefinition(PageIdDeterminator::class)) {
             return;
         }