+++ src/Request/ContaoPageContextInitializer.php
@@ -1,15 +1,5 @@
 <?php
 
-/**
- * Contao Page Context
- *
- * @package    contao-page-context
- * @author     David Molineus <david.molineus@netzmacht.de>
- * @copyright  2018 netzmacht David Molineus.
- * @license    LGPL-3.0 https://github.com/netzmacht/contao-page-context/blob/master/LICENSE
- * @filesource
- */
-
 declare(strict_types=1);
 
 namespace Netzmacht\Contao\PageContext\Request;
@@ -17,7 +7,7 @@ namespace Netzmacht\Contao\PageContext\Request;
 use Contao\Config;
 use Contao\Controller;
 use Contao\CoreBundle\Exception\NoLayoutSpecifiedException;
-use Contao\CoreBundle\Framework\ContaoFrameworkInterface;
+use Contao\CoreBundle\Framework\ContaoFramework;
 use Contao\CoreBundle\Image\PictureFactoryInterface;
 use Contao\CoreBundle\Monolog\ContaoContext;
 use Contao\LayoutModel;
@@ -26,15 +16,20 @@ use Contao\PageRegular;
 use Contao\StringUtil;
 use Contao\System;
 use Contao\ThemeModel;
-use Netzmacht\Contao\Toolkit\Callback\Invoker;
 use Netzmacht\Contao\Toolkit\Data\Model\RepositoryManager;
 use Psr\Log\LoggerInterface;
 use Psr\Log\LogLevel;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\Translation\TranslatorInterface;
+use Symfony\Contracts\Translation\LocaleAwareInterface;
+
+use function array_merge;
+use function assert;
 use function define;
 use function defined;
+use function explode;
+use function is_array;
 use function is_string;
+use function str_replace;
 
 /**
  * Class PageContextInitializer initialize the page context which is usually done by the Contao regular page.
@@ -44,7 +39,7 @@ final class ContaoPageContextInitializer implements PageContextInitializer
     /**
      * Default config.
      *
-     * @var array
+     * @var array<string,bool>
      */
     private $defaults = [
         'BE_USER_LOGGED_IN' => false,
@@ -54,23 +49,16 @@ final class ContaoPageContextInitializer implements PageContextInitializer
     /**
      * Translator.
      *
-     * @var TranslatorInterface
+     * @var LocaleAwareInterface
      */
     private $translator;
 
     /**
      * Contao framework.
      *
-     * @var ContaoFrameworkInterface
+     * @var ContaoFramework
      */
     private $framework;
-
-    /**
-     * Callback invoker.
-     *
-     * @var Invoker
-     */
-    private $callbackInvoker;
 
     /**
      * Picture factory.
@@ -94,20 +82,16 @@ final class ContaoPageContextInitializer implements PageContextInitializer
     private $logger;
 
     /**
-     * PageContextInitializer constructor.
-     *
-     * @param TranslatorInterface      $translator        Translator.
-     * @param ContaoFrameworkInterface $framework         Contao framework.
-     * @param Invoker                  $callbackInvoker   Callback invoker.
-     * @param PictureFactoryInterface  $pictureFactory    Picture factory.
-     * @param RepositoryManager        $repositoryManager Repository manager.
-     * @param LoggerInterface          $logger            Logger.
-     * @param array                    $defaults          Default config to override default configs.
+     * @param LocaleAwareInterface    $translator        Translator.
+     * @param ContaoFramework         $framework         Contao framework.
+     * @param PictureFactoryInterface $pictureFactory    Picture factory.
+     * @param RepositoryManager       $repositoryManager Repository manager.
+     * @param LoggerInterface         $logger            Logger.
+     * @param array<string,bool>      $defaults          Default config to override default configs.
      */
     public function __construct(
-        TranslatorInterface $translator,
-        ContaoFrameworkInterface $framework,
-        Invoker $callbackInvoker,
+        LocaleAwareInterface $translator,
+        ContaoFramework $framework,
         PictureFactoryInterface $pictureFactory,
         RepositoryManager $repositoryManager,
         LoggerInterface $logger,
@@ -115,7 +99,6 @@ final class ContaoPageContextInitializer implements PageContextInitializer
     ) {
         $this->translator        = $translator;
         $this->framework         = $framework;
-        $this->callbackInvoker   = $callbackInvoker;
         $this->pictureFactory    = $pictureFactory;
         $this->repositoryManager = $repositoryManager;
         $this->logger            = $logger;
@@ -127,8 +110,6 @@ final class ContaoPageContextInitializer implements PageContextInitializer
      *
      * @param PageContext $context Page context.
      * @param Request     $request Web request.
-     *
-     * @return void
      */
     public function initialize(PageContext $context, Request $request): void
     {
@@ -137,25 +118,25 @@ final class ContaoPageContextInitializer implements PageContextInitializer
         $this->initializeGlobals($context);
         $this->initializeLocale($context, $request);
         $this->initializeStaticUrls();
-        $this->initializePageLayout($context, $request);
+        $this->initializePageLayout($context);
     }
 
     /**
      * Initialize user logged in constants set by default.
      *
      * You can't trust this constants, as only defaults values are set right now.
-     *
-     * @return void
      */
     private function initializeUserLoggedInConstants(): void
     {
-        if (!defined('BE_USER_LOGGED_IN')) {
+        if (! defined('BE_USER_LOGGED_IN')) {
             define('BE_USER_LOGGED_IN', $this->defaults['BE_USER_LOGGED_IN']);
         }
 
-        if (!defined('FE_USER_LOGGED_IN')) {
-            define('FE_USER_LOGGED_IN', $this->defaults['FE_USER_LOGGED_IN']);
+        if (defined('FE_USER_LOGGED_IN')) {
+            return;
         }
+
+        define('FE_USER_LOGGED_IN', $this->defaults['FE_USER_LOGGED_IN']);
     }
 
     /**
@@ -163,15 +144,13 @@ final class ContaoPageContextInitializer implements PageContextInitializer
      *
      * @param PageContext $context The page context.
      *
-     * @return void
-     *
      * @SuppressWarnings(PHPMD.Superglobals)
      */
     private function initializeGlobals(PageContext $context): void
     {
         $page = $context->page();
 
-        if ($page->adminEmail != '') {
+        if ($page->adminEmail !== '') {
             $adminEmail = $page->adminEmail;
         } else {
             $adminEmail = $this->framework->getAdapter(Config::class)->get('adminEmail');
@@ -189,8 +168,6 @@ final class ContaoPageContextInitializer implements PageContextInitializer
      *
      * @param PageContext $context Page context.
      * @param Request     $request Web request.
-     *
-     * @return void
      */
     private function initializeLocale(PageContext $context, Request $request): void
     {
@@ -204,8 +181,6 @@ final class ContaoPageContextInitializer implements PageContextInitializer
 
     /**
      * Initialize static urls.
-     *
-     * @return void
      */
     private function initializeStaticUrls(): void
     {
@@ -216,24 +191,25 @@ final class ContaoPageContextInitializer implements PageContextInitializer
      * Initialize the page layout.
      *
      * @param PageContext $context Page context.
-     * @param Request     $request Web request.
-     *
-     * @return void
      *
      * @SuppressWarnings(PHPMD.Superglobals)
      */
-    private function initializePageLayout(PageContext $context, Request $request): void
+    private function initializePageLayout(PageContext $context): void
     {
         $page        = $context->page();
-        $layout      = $this->getPageLayout($page, $request);
+        $layout      = $this->getPageLayout($page);
         $pageRegular = new PageRegular();
 
-        if (isset($GLOBALS['TL_HOOKS']['getPageLayout']) && \is_array($GLOBALS['TL_HOOKS']['getPageLayout'])) {
-            $this->callbackInvoker->invokeAll($GLOBALS['TL_HOOKS']['getPageLayout'], [$page, $layout, $pageRegular]);
+        if (isset($GLOBALS['TL_HOOKS']['getPageLayout']) && is_array($GLOBALS['TL_HOOKS']['getPageLayout'])) {
+            $systemAdapter = $this->framework->getAdapter(System::class);
+            foreach ($GLOBALS['TL_HOOKS']['generatePage'] as $callback) {
+                $callback[0] = $systemAdapter->__call('importStatic', [$callback[0]]);
+                $callback[0]->{$callback[1]}($page, $layout, $pageRegular);
+            }
         }
 
-        /** @var ThemeModel $theme */
         $theme = $this->repositoryManager->getRepository(ThemeModel::class)->find((int) $layout->pid);
+        assert($theme instanceof ThemeModel);
 
         // Set the default image densities
         $this->pictureFactory->setDefaultDensities($theme->defaultImageDensities);
@@ -246,45 +222,42 @@ final class ContaoPageContextInitializer implements PageContextInitializer
         $page->templateGroup = $theme->templates;
 
         $doctype = $layout->doctype;
-        if (is_string($doctype)) {
-            // Store the output format
-            [$strFormat, $strVariant] = explode('_', $doctype);
-
-            $page->outputFormat  = $strFormat;
-            $page->outputVariant = $strVariant;
+        if (! is_string($doctype)) {
+            return;
         }
+
+        // Store the output format
+        [$strFormat, $strVariant] = explode('_', $doctype);
+
+        $page->outputFormat  = $strFormat;
+        $page->outputVariant = $strVariant;
     }
 
     /**
      * Get a page layout and return it as database result object
      *
      * @param PageModel $pageModel The page model.
-     * @param Request   $request   Web request.
-     *
-     * @return LayoutModel
      *
      * @throws NoLayoutSpecifiedException If no page layout could be found.
      */
-    private function getPageLayout(PageModel $pageModel, Request $request): LayoutModel
+    private function getPageLayout(PageModel $pageModel): LayoutModel
     {
-        /** @var LayoutModel $layoutModel */
-        $isMobile    = $request->query->get('view') === 'mobile';
-        $layoutId    = (int) (($isMobile && $pageModel->mobileLayout) ? $pageModel->mobileLayout : $pageModel->layout);
+        $layoutId    = (int) $pageModel->layout;
         $layoutModel = $this->repositoryManager->getRepository(LayoutModel::class)->find($layoutId);
 
         // Die if there is no layout
-        if ($layoutModel === null) {
+        if (! $layoutModel instanceof LayoutModel) {
             $this->logger->log(
                 LogLevel::ERROR,
                 'Could not find layout ID "' . $layoutId . '"',
                 ['contao' => new ContaoContext(__METHOD__, LogLevel::ERROR)]
             );
+
             throw new NoLayoutSpecifiedException('No layout specified');
         }
 
         $pageModel->hasJQuery   = $layoutModel->addJQuery;
         $pageModel->hasMooTools = $layoutModel->addMooTools;
-        $pageModel->isMobile    = $isMobile;
 
         return $layoutModel;
     }