+++ src/Request/PageContext.php
@@ -1,24 +1,11 @@
 <?php
 
-/**
- * Contao Page Context
- *
- * @package    contao-page-context
- * @author     David Molineus <david.molineus@netzmacht.de>
- * @copyright  2018 netzmacht David Molineus.
- * @license    LGPL-3.0 https://github.com/netzmacht/contao-page-context/blob/master/LICENSE
- * @filesource
- */
-
 declare(strict_types=1);
 
 namespace Netzmacht\Contao\PageContext\Request;
 
 use Contao\PageModel;
 
-/**
- * Class PageContext describes a current page
- */
 final class PageContext
 {
     /**
@@ -36,8 +23,6 @@ final class PageContext
     private $rootPage;
 
     /**
-     * PageContext constructor.
-     *
      * @param PageModel $pageModel Page model.
      * @param PageModel $rootPage  Root page model.
      */
@@ -49,8 +34,6 @@ final class PageContext
 
     /**
      * Get the page model.
-     *
-     * @return PageModel
      */
     public function page(): PageModel
     {
@@ -59,8 +42,6 @@ final class PageContext
 
     /**
      * Get the root page model.
-     *
-     * @return PageModel
      */
     public function rootPage(): PageModel
     {