+++ src/Request/PageContextFactory.php
@@ -1,15 +1,5 @@
 <?php
 
-/**
- * Contao Page Context
- *
- * @package    contao-page-context
- * @author     David Molineus <david.molineus@netzmacht.de>
- * @copyright  2018 netzmacht David Molineus.
- * @license    LGPL-3.0 https://github.com/netzmacht/contao-page-context/blob/master/LICENSE
- * @filesource
- */
-
 declare(strict_types=1);
 
 namespace Netzmacht\Contao\PageContext\Request;
@@ -18,9 +8,6 @@ use Contao\PageModel;
 use Netzmacht\Contao\PageContext\Exception\InitializePageContextFailed;
 use Netzmacht\Contao\Toolkit\Data\Model\RepositoryManager;
 
-/**
- * Class PageContextFactory creates the page context from a request or page id
- */
 final class PageContextFactory
 {
     /**
@@ -31,8 +18,6 @@ final class PageContextFactory
     private $repositoryManager;
 
     /**
-     * PageContextFactory constructor.
-     *
      * @param RepositoryManager $repositoryManager Repository manager.
      */
     public function __construct(RepositoryManager $repositoryManager)
@@ -45,8 +30,6 @@ final class PageContextFactory
      *
      * @param int $pageId Page id for the page context.
      *
-     * @return PageContext
-     *
      * @throws InitializePageContextFailed When creating page context failed.
      */
     public function __invoke(int $pageId): PageContext
@@ -54,15 +37,16 @@ final class PageContextFactory
         $repository = $this->repositoryManager->getRepository(PageModel::class);
         $pageModel  = $repository->find($pageId);
 
-        if (!$pageModel instanceof PageModel) {
+        if (! $pageModel instanceof PageModel) {
             throw InitializePageContextFailed::pageNotFound($pageId);
         }
 
         $pageModel->loadDetails();
+        /** @psalm-suppress RedundantCastGivenDocblockType - Contao doc type issue */
         $rootPage = $repository->find((int) $pageModel->rootId);
 
-        if (!$rootPage instanceof PageModel) {
-            throw InitializePageContextFailed::rootPageNotFound($pageModel->id);
+        if (! $rootPage instanceof PageModel) {
+            throw InitializePageContextFailed::rootPageNotFound((int) $pageModel->id);
         }
 
         return new PageContext($pageModel, $rootPage);